calendar cleanup with helpers
