+++ src/components/calendar/Calendar.jsx
@@ -3,6 +3,7 @@ import './Calendar.css';
 import FullCalendar from '../fullCalendar/FullCalendar.jsx';
 import questionmark from "../../assets/question.png";
 import ExplenationMessage from '../explenationMessage/ExplenationMessage.jsx';
+import { getNextSixDays, getPhaseForDate, formatDate } from '../../helpers/calendarHelpers.jsx';
 
 const Calendar = ({ onPhaseChange }) => {
     const [currentDate, setCurrentDate] = useState(new Date());
@@ -10,27 +11,9 @@ const Calendar = ({ onPhaseChange }) => {
     const [phases, setPhases] = useState([]);
     const [showExplenation, setShowExplenation] = useState(false);
 
-    const getNextSixDays = (date) => {
-        const dates = [];
-        for (let i = 1; i <= 6; i++) {
-            const newDate = new Date(date);
-            newDate.setDate(date.getDate() + i);
-            dates.push(newDate);
-        }
-        return dates;
-    };
-
-    const handlePrevDay = () => {
-        const newDate = new Date(currentDate.setDate(currentDate.getDate() - 1));
-        setCurrentDate(newDate);
-        updateCurrentPhase(newDate);
-    };
-
-    const handleNextDay = () => {
-        const newDate = new Date(currentDate.setDate(currentDate.getDate() + 1));
-        setCurrentDate(newDate);
-        updateCurrentPhase(newDate);
-    };
+    useEffect(() => {
+        fetchCycle();
+    }, []);
 
     const fetchCycle = async () => {
         try {
@@ -52,39 +35,25 @@ const Calendar = ({ onPhaseChange }) => {
         }
     };
 
-    useEffect(() => {
-        fetchCycle();
-    }, []);
-
-    const getPhaseForDate = (date) => {
-        const startOfDay = (date) => new Date(date.getFullYear(), date.getMonth(), date.getDate());
-
-        const targetDate = startOfDay(date);
-
-        return phases.find(phase => {
-            const startDate = startOfDay(new Date(phase.startDate));
-            const endDate = new Date(new Date(phase.endDate).setHours(23, 59, 59, 999));
-
-            return targetDate >= startDate && targetDate <= endDate;
-        });
-    };
-
     const updateCurrentPhase = (date) => {
-        const phase = getPhaseForDate(date);
-        if (phase) {
-            onPhaseChange(phase.phaseName);
-        } else {
-            onPhaseChange("No Phase");
-        }
+        const phase = getPhaseForDate(date, phases);
+        onPhaseChange(phase ? phase.phaseName : "No Phase");
     };
 
-    const formatDate = (date) => {
-        return date.toLocaleString('en-US', { weekday: 'short' }).toUpperCase();
+    const handlePrevDay = () => {
+        const newDate = new Date(currentDate.setDate(currentDate.getDate() - 1));
+        setCurrentDate(newDate);
+        updateCurrentPhase(newDate);
+    };
+
+    const handleNextDay = () => {
+        const newDate = new Date(currentDate.setDate(currentDate.getDate() + 1));
+        setCurrentDate(newDate);
+        updateCurrentPhase(newDate);
     };
 
     const nextSixDays = getNextSixDays(currentDate);
-
-    const currentPhase = getPhaseForDate(currentDate);
+    const currentPhase = getPhaseForDate(currentDate, phases);
     const title = currentPhase ? currentPhase.phaseName : (phases.length === 0 ? "Sync your Cycle" : "No Phase");
     const phaseClass = currentPhase ? currentPhase.phaseName.toLowerCase() : "";
 
@@ -104,7 +73,7 @@ const Calendar = ({ onPhaseChange }) => {
             </div>
             <div className="calendar-grid">
                 {nextSixDays.map((date, index) => {
-                    const phase = getPhaseForDate(date);
+                    const phase = getPhaseForDate(date, phases);
                     let className = "calendar-day";
                     if (phase) {
                         className += ` ${phase.phaseName.toLowerCase()}`;
